[WIP] (Re-)Create the PR template file and update the text.
Small PR to add the existing PR template to this repo. Currently this
template is checked into alot of repositories as a file which makes
editing the file slightly annoying as it must be edited everywhere.

This PR propoes to submodule the pr template and then have repositories
pull in the latest version of the template.

Again this change must be made everywhere initially but afterwards
updates to the dorfr will only have to be made in one repository.