[CES-3227] Ignore whitespacing after commas
When executing `pycodestyle -h`, the default list of ignored tests
is:
```
  --ignore=errors      skip errors and warnings (e.g. E4,W) (default:
                       E121,E123,E126,E226,E24,E704,W503,W504)
```
It appears that because run_pycodestyle.sh  specifies a --config with
an ignore parameter, the default set of warnings/errors is overridden
to be:
```
ignore = E722, W503, E266, E501
```
This then enables the E24 family of errors, to which E241 belongs.

Further review of the python3.11/site-packages/pycodestyle.py
indicates that, although there is support for noqa, not all checks
have support implemented; contrast:
```
@register_check
def whitespace_around_comma(logical_line):
    r"""Avoid extraneous whitespace after a comma or a colon.

    Note: these checks are disabled by default

SNIP
```

and:

```
@register_check
def continued_indentation(logical_line, tokens, indent_level, hang_closing,
                          indent_char, indent_size, noqa, verbose):
    r"""Continuation lines indentation.

SNIP
```

There also does not seem to be a solution for ignoring files altogether
using noqa.

This mean that realistically the only solution is to add E241 to the
ignore list.